finished components split
